+++ crates/app/src/commands/setup.rs
@@ -8,7 +8,7 @@ use tracing::instrument;
 pub async fn setup(session: CliSession) -> AppResult {
     let done = create_progress_bar("Downloading and installing tools...");
 
-    analyze::load_toolchain(session.get_toolchain_registry()?).await?;
+    analyze::load_toolchain(session.get_toolchain_registry().await?).await?;
 
     done("Setup complete", true);