+++ crates/toolchain-plugin/src/toolchain_registry.rs
@@ -1,9 +1,7 @@
 use crate::toolchain_plugin::ToolchainPlugin;
 use miette::IntoDiagnostic;
-use moon_config::ToolConfig;
-use moon_plugin::{
-    serialize_config, MoonEnvironment, PluginId, PluginRegistry, PluginType, ProtoEnvironment,
-};
+use moon_config::ToolchainPluginConfig;
+use moon_plugin::{serialize_config, PluginHostData, PluginId, PluginRegistry, PluginType};
 use proto_core::inject_proto_manifest_config;
 use rustc_hash::FxHashMap;
 use std::ops::Deref;
@@ -13,24 +11,28 @@ use tracing::{debug, trace};
 
 #[derive(Debug)]
 pub struct ToolchainRegistry {
-    pub configs: FxHashMap<PluginId, ToolConfig>,
+    pub configs: FxHashMap<PluginId, ToolchainPluginConfig>,
     registry: Arc<PluginRegistry<ToolchainPlugin>>,
 }
 
 impl ToolchainRegistry {
-    pub fn new(moon_env: Arc<MoonEnvironment>, proto_env: Arc<ProtoEnvironment>) -> Self {
+    pub fn new(host_data: PluginHostData) -> Self {
         Self {
             configs: FxHashMap::default(),
-            registry: Arc::new(PluginRegistry::new(
-                PluginType::Toolchain,
-                moon_env,
-                proto_env,
-            )),
+            registry: Arc::new(PluginRegistry::new(PluginType::Toolchain, host_data)),
         }
     }
 
+    pub fn get_plugin_ids(&self) -> Vec<&PluginId> {
+        self.configs.keys().collect()
+    }
+
+    pub fn has_plugins(&self) -> bool {
+        !self.configs.is_empty()
+    }
+
     pub async fn load_all(&self) -> miette::Result<()> {
-        if self.configs.is_empty() {
+        if !self.has_plugins() {
             return Ok(());
         }
 
@@ -56,7 +58,7 @@ impl ToolchainRegistry {
                             .config
                             .insert("moon_toolchain_config".to_owned(), value);
 
-                        inject_proto_manifest_config(&id, &registry.proto_env, manifest)?;
+                        inject_proto_manifest_config(&id, &registry.host_data.proto_env, manifest)?;
 
                         Ok(())
                     })